I'm adding a few changes here to port the hadoop plugin to 0.18, along with a few more file renames per Doug's advice.  From Doug:
I recommend going back to hadoop-plugin.xml for the name though.
One problem is on the HQ server-side, resource types are tied to the
plugin name in the database so it's best to leave it as 'hadoop'.  It'll
also be much easier for me to merge your changes.  Maybe just change the
classnames to 0.18's flavor for now and I'll add some groovy scripting
later to check for both flavors.
There is another option to version the resource types in the plugin.xml,
but that can ugly to keep up with.